refactor: extract processor to own module
